fixing group tests
