+++ .gitignore
@@ -7,3 +7,4 @@ composer.phar
 .env.php
 .DS_Store
 Thumbs.db
+composer.lock
