+++ composer.json
@@ -1,0 +1,24 @@
+{
+    "name": "mgp25/libaxolotl-php",
+    "type": "library",
+    "description": "This is a php port of libaxolotl-android originally written by Moxie Marlinspike",
+    "require": {
+        "php": ">= 5.6",
+        "ext-curve25519": "*",
+        "ext-protobuf": "*",
+        "ext-crypto": "*"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "~4.8"
+    },
+    "autoload": {
+        "psr-4": {
+            "Libaxolotl\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Libaxolotl\\Tests\\": "tests/"
+        }
+    }
+}
