+++ src/ecc/Curve.php
@@ -1,11 +1,14 @@
 <?php
+namespace Libaxolotl\ecc;
+
+use Libaxolotl\ecc\ECKeyPair;
+use Libaxolotl\ecc\DjbECPublicKey;
+use Libaxolotl\ecc\ECPublicKey;
+use Libaxolotl\ecc\ECPrivateKey;
+use Libaxolotl\ecc\DjbECPrivateKey;
 
 require_once __DIR__.'/../InvalidKeyException.php';
-require_once __DIR__.'/ECKeyPair.php';
-require_once __DIR__.'/ECPrivateKey.php';
-require_once __DIR__.'/ECPublicKey.php';
-require_once __DIR__.'/DjbECPublicKey.php';
-require_once __DIR__.'/DjbECPrivateKey.php';
+
 class Curve
 {
     const DJB_TYPE = 0x05;  // int
