+++ src/ecc/DjbECPrivateKey.php
@@ -1,10 +1,11 @@
 <?php
+namespace Libaxolotl\ecc;
 
 class DjbECPrivateKey implements ECPrivateKey
 {
     protected $privateKey;    // byte[] --> php string now
 
-    public function DjbECPrivateKey($privateKey) // [byte[] privateKey]
+    public function __construct($privateKey) // [byte[] privateKey]
     {
         $this->privateKey = $privateKey;
     }
