+++ src/protocol/PreKeyWhisperMessage.php
@@ -1,8 +1,10 @@
 <?php
+
+use Libaxolotl\protocol\CiphertextMessage;
+
     require_once __DIR__.'/../ecc/ECPublicKey.php';
     require_once __DIR__.'/../util/ByteUtil.php';
     require_once __DIR__.'/pb_proto_WhisperTextProtocol.php';
-    require_once __DIR__.'/CiphertextMessage.php';
     require_once __DIR__.'/../InvalidMessageException.php';
     require_once __DIR__.'/../InvalidVersionException.php';
     require_once __DIR__.'/../LegacyMessageException.php';
@@ -19,7 +21,7 @@
         protected $message;
         protected $serialized;
 
-        public function PreKeyWhisperMessage($messageVersion = null,
+        public function __construct($messageVersion = null,
                                              $registrationId = null,
                                              $preKeyId = null,
                                              $signedPreKeyId = null,
