+++ src/protocol/SenderKeyMessage.php
@@ -1,8 +1,10 @@
 <?php
+
+use Libaxolotl\protocol\CiphertextMessage;
+
     require_once __DIR__.'/../ecc/ECPublicKey.php';
     require_once __DIR__.'/../util/ByteUtil.php';
     require_once __DIR__.'/pb_proto_WhisperTextProtocol.php';
-    require_once __DIR__.'/CiphertextMessage.php';
     require_once __DIR__.'/../InvalidMessageException.php';
     require_once __DIR__.'/../LegacyMessageException.php';
     class SenderKeyMessage extends CiphertextMessage
@@ -14,7 +16,7 @@
         protected $iteration;
         protected $ciphertext;
 
-        public function SenderKeyMessage($keyId = null, $iteration = null, $ciphertext = null, $signatureKey = null, $serialized = null)
+        public function __construct($keyId = null, $iteration = null, $ciphertext = null, $signatureKey = null, $serialized = null)
         {
             if ($serialized == null) {
                 $version = ByteUtil::intsToByteHighAndLow(self::CURRENT_VERSION, self::CURRENT_VERSION);
