+++ src/protocol/WhisperMessage.php
@@ -1,8 +1,10 @@
 <?php
+
+use Libaxolotl\protocol\CiphertextMessage;
+
     require_once __DIR__.'/../ecc/ECPublicKey.php';
     require_once __DIR__.'/../util/ByteUtil.php';
     require_once __DIR__.'/pb_proto_WhisperTextProtocol.php';
-    require_once __DIR__.'/CiphertextMessage.php';
     require_once __DIR__.'/../InvalidMessageException.php';
     require_once __DIR__.'/../LegacyMessageException.php';
     class WhisperMessage extends CiphertextMessage
@@ -15,7 +17,7 @@
         protected $cipherText;
         protected $serialized;
 
-        public function WhisperMessage($messageVersion = null,
+        public function __construct($messageVersion = null,
                                         $macKey = null,
                                         $senderRatchetKey = null,
                                         $counter = null,
