+++ tests/inmemoryaxolotlstore.php
