+++ tests/inmemoryprekeystore.php
