+++ tests/ratchet/test_rootkey.php
