+++ tests/test_sessionbuilder.php
