+++ tests/test_sessioncipher.php
