+++ tests/test_sigs.php
