+++ tests/util/test_byteutil.php
